+++ ClipboardArchivio/App/ClipboardArchivioApp.swift
@@ -25,17 +25,11 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     let appState = AppState()
     private(set) var statusBarController: StatusBarController?
     private let preferencesWindow = PreferencesWindowController()
-    private let onboardingWindow = OnboardingWindowController()
     private var localKeyboardMonitor: Any?
     private var globalKeyboardMonitor: Any?
 
     func applicationDidFinishLaunching(_ notification: Notification) {
-        if OnboardingStore.isCompleted {
-            NSApp.setActivationPolicy(.accessory)
-        } else {
-            NSApp.setActivationPolicy(.regular)
-            NSApp.activate(ignoringOtherApps: true)
-        }
+        NSApp.setActivationPolicy(.accessory)
 
         preferencesWindow.configure(appState: appState)
 
@@ -45,18 +39,9 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
 
         statusBarController = StatusBarController(appState: appState)
 
-        onboardingWindow.configure(appState: appState)
-        onboardingWindow.onFinish = { [weak self] in
-            self?.finishFirstLaunchOnboarding()
-        }
-
         appState.historyStore.refreshExpiredItems()
         appState.launchAtLogin.applyDefaultIfNeeded()
         setupKeyboardShortcuts()
-
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.35) { [weak self] in
-            self?.onboardingWindow.showIfNeeded()
-        }
     }
 
     var isClipboardPanelVisible: Bool {
@@ -66,11 +51,6 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     func showPreferences() {
         statusBarController?.closePanelIfVisible()
         preferencesWindow.show(relativeTo: statusBarController?.clipboardPanelWindow)
-    }
-
-    private func finishFirstLaunchOnboarding() {
-        NSApp.setActivationPolicy(.accessory)
-        statusBarController?.openArchivePanel()
     }
 
     func applicationShouldTerminateAfterLastWindowClosed(_ sender: NSApplication) -> Bool {